third_party/blink/web_tests/platform/generic/fast/regex/cross-frame-callable-expected.txt

PASS re('a') threw exception TypeError: re is not a function.
PASS re.exec('a') is ['a']
PASS successfullyParsed is true

TEST COMPLETE

DONE
